Updated extension dependencies
We want to support 3.0 with this PR. Also, it's using some code
from `solidus_support` that was introduced in 0.8.1, so this
should be the minimum version required. Set to 0.8 to losen restrictions.
Webdrivers has been added to support capybara for specs. Bumped
Deface to 1.5 to keep in line with Solidus. Removed duplicated rubocop
gems that are provided by SolidusDevSupport